critical path on swap works
