+++ scripts/redrawBountyThreads.js
@@ -1,0 +1,1 @@
+//TODONOW finish
