+++ source/frontend/context_menus/Record_Bounty_Turn-In.js
@@ -1,7 +1,7 @@
 const { InteractionContextType, PermissionFlagsBits, ModalBuilder, ActionRowBuilder, TextInputBuilder, TextInputStyle, userMention, bold, MessageFlags, DiscordjsErrorCodes } = require('discord.js');
 const { UserContextMenuWrapper } = require('../classes');
 const { SKIP_INTERACTION_HANDLING } = require('../../constants');
-const { commandMention, congratulationBuilder, buildBountyEmbed, generateBountyBoardButtons } = require('../shared');
+const { commandMention, congratulationBuilder, buildBountyEmbed } = require('../shared');
 
 /** @type {typeof import("../../logic")} */
 let logicLayer;
@@ -54,10 +54,7 @@ module.exports = new UserContextMenuWrapper(mainId, PermissionFlagsBits.SendMess
 				await thread.setArchived(false, "Unarchived to update posting");
 			}
 			post.send({ content: `${userMention(interaction.targetId)} has turned-in this bounty! ${congratulationBuilder()}!` });
-			(await post.fetchStarterMessage()).edit({
-				embeds: [await buildBountyEmbed(bounty, modalSubmission.guild, origin.hunter.getLevel(origin.company.xpCoefficient), false, origin.company, new Set([interaction.targetId]))],
-				components: generateBountyBoardButtons(bounty)
-			});
+			(await post.fetchStarterMessage()).edit({ embeds: [await buildBountyEmbed(bounty, modalSubmission.guild, origin.hunter.getLevel(origin.company.xpCoefficient), false, origin.company, new Set([interaction.targetId]))] });
 			modalSubmission.reply({ content: `${userMention(interaction.targetId)}'s turn-in of ${bold(bounty.title)} has been recorded! They will recieve the reward XP when you ${commandMention("bounty complete")}.`, flags: MessageFlags.Ephemeral });
 		}).catch(error => {
 			if (error.code !== DiscordjsErrorCodes.InteractionCollectorError) {
